% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/macgrubeR.R
\name{macgrubeR}
\alias{macgrubeR}
\title{Print a line from the feature film MacGruber}
\usage{
macgrubeR()
}
\arguments{
\item{There}{are no function arguments}
}
\value{
There is no return value
}
\description{
Print a line from the feature film MacGruber
}
\examples{
macgrubeR()
}

